actualizar servCrearSolicitud

+++ proyecto-gestion-solicitudes-main/src/main/java/com/campus360/solicitudes/DTOs/GenerarOrdenPagoRequest.java
@@ -4,15 +4,15 @@ import java.math.BigDecimal;
 
 public class GenerarOrdenPagoRequest {
 
-    private Long solicitudId;
+    private Integer solicitudId;
     private BigDecimal monto;
 
     // getters y setters
-    public Long getSolicitudId() {
+    public Integer getSolicitudId() {
         return solicitudId;
     }
 
-    public void setSolicitudId(Long solicitudId) {
+    public void setSolicitudId(Integer solicitudId) {
         this.solicitudId = solicitudId;
     }
